Refactored enums out into enums folder.
