+++ apps/augury-backend/src/config/enums/PortfolioColor.ts
@@ -1,0 +1,7 @@
+export enum PortfolioColor {
+  WHITE = 'white',
+  RED = 'red',
+  GREEN = 'green',
+  BLUE = 'blue',
+  YELLOW = 'yellow',
+}
