+++ apps/augury-backend/src/config/enums/PortfolioRisk.ts
@@ -1,0 +1,5 @@
+export enum PortfolioRisk {
+  CONSERVATIVE = 'conservative',
+  MODERATE = 'moderate',
+  AGGRESSIVE = 'aggressive',
+}
